+++ StockTracker.FrontEnd/src/components/Clients/Form.vue
